# internal/domain/dto/member/addmember.go
package dto

// AddMemberRequest Model of CreateMemberRequestHandler
type AddMemberRequest struct {
	Name     string `json:"name" validate:"required"`
	Email    string `json:"email" validate:"required"`
	Password string `json:"password" validate:"required"`
}